+++ Changelog.md
@@ -13,6 +13,7 @@ This file exists because sometimes I wait to commit my changes until I've done a
 - Add `ignore_prefixes` config option
 - Save monthly stats in addition to all-time stats
 - Beef up stats command with embeds and option to check another user's stats
+- Allow placeholders to be used in the system prompt
 
 # 2023-03-17
 - Change config structure